add album art, restructure elements and track data into JSON
